+++ services/music-service/src/MusicService.Application/Interfaces/IEventPublisher.cs
@@ -1,0 +1,10 @@
+using System.Threading;
+using System.Threading.Tasks;
+using MusicService.Domain.Events;
+
+namespace MusicService.Application.Interfaces;
+
+public interface IEventPublisher
+{
+    Task PublishNewReleaseAsync(NewReleaseEvent @event, CancellationToken cancellationToken = default);
+}